movies/views/movie.py
from rest_framework import viewsets

from movies import models, serializers


class MovieViewSet(viewsets.ModelViewSet):
    queryset = models.Movie.objects.all()
    serializer_class = serializers.MovieSerializer
